requirements.txt
pyyaml==5.4.1
selenium==3.141.0
PyPDF2==1.26.0
pwinput==1.0.2
dacite==1.7.0
pandas==1.5.3
openpyxl==3.1.0